requirements.txt
numpy==1.24.2
opencv-contrib-python==4.7.0.72
pyserial==3.5
